All Tasks is are over.

+++ MusicApp/MusicApp/Controllers/SongController.cs
@@ -1,7 +1,9 @@
 ﻿using DataBase.Models;
+using FluentValidation;
 using Microsoft.AspNetCore.Mvc;
 using MusicApp.Contracts;
 using MusicApp.Services;
+using MusicApp.Validations;
 
 namespace MusicApp.Controllers;
 
@@ -13,23 +15,27 @@ public class SongsController : ControllerBase
     public ActionResult<List<SongContract>> GetAllSongsForAlbum(string artistId, string albumId)
     {
         var album = AlbumService.GetAlbumById(albumId);
+        var artist = ArtistService.GetArtistById(artistId);
 
-        if (album == null)
+        if (artist is null || album is null)
         {
-            return NotFound();
-        }
-
-        return Ok(SongService.GetAllSongsForAlbum(albumId));
+            return NotFound("ArtistId or AlbumId not found");
+        } else
+        {
+            return Ok(SongService.GetAllSongsForAlbum(albumId));
+        }        
     }
 
     [HttpGet("{id}")]
     public ActionResult<SongContract> GetSongById(string artistId, string albumId, string id)
     {
         var song = SongService.GetSongById(id);
+        var artist = ArtistService.GetArtistById(artistId);
+        var album = AlbumService.GetAlbumById(albumId);
 
-        if (song is null)
+        if (artist is null || album is null ||song is null)
         {
-            return NotFound();
+            return NotFound("Artist or Album or Song is not found");
         }
 
         return Ok(song);
@@ -38,6 +44,26 @@ public class SongsController : ControllerBase
     [HttpPost]
     public IActionResult CreateSongs(string artistId, string albumId, List<SongCreateContract> songs)
     {
+        var validator = new SongsCreateValidation();
+        var result = validator.Validate(songs);
+        if (!result.IsValid)
+        {
+            var messages = result.Errors.Select((e) =>
+            {
+                return new Error
+                {
+                    Message = e.ErrorMessage,
+                    Field = e.PropertyName
+                };
+            });
+
+            var errors = new Errors
+            {
+                ErrorsMessages = messages.ToList(),
+            };
+            return BadRequest(errors);
+        }
+
         var album = AlbumService.GetAlbumById(albumId);
 
         if (album == null)
@@ -53,12 +79,32 @@ public class SongsController : ControllerBase
     [HttpPut]
     public IActionResult UpdateSongs(string artistId, string albumId, List<SongUpdateContract> songs)
     {
+        var validator = new SongsUpdateValidation();
+        var result = validator.Validate(songs);
+        if (!result.IsValid)
+        {
+            var messages = result.Errors.Select((e) =>
+            {
+                return new Error
+                {
+                    Message = e.ErrorMessage,
+                    Field = e.PropertyName
+                };
+            });
+
+            var errors = new Errors
+            {
+                ErrorsMessages = messages.ToList(),
+            };
+            return BadRequest(errors);
+        }
+
         var artist = ArtistService.GetArtistById(artistId);
         var album = AlbumService.GetAlbumById(albumId);
 
         if (artist == null || album == null)
         {
-            return NotFound();
+            return NotFound("Artist or Album is not found");
         }
 
         SongService.UpdateSongs(songs);
@@ -69,11 +115,33 @@ public class SongsController : ControllerBase
     [HttpPut("{id}")]
     public IActionResult UpdateSong(string artistId, string albumId,  string id, SongUpdateContract song)
     {
+        var validator = new SongUpdateValidation();
+        var result = validator.Validate(song);
+        if (!result.IsValid)
+        {
+            var messages = result.Errors.Select((e) =>
+            {
+                return new Error
+                {
+                    Message = e.ErrorMessage,
+                    Field = e.PropertyName
+                };
+            });
+
+            var errors = new Errors
+            {
+                ErrorsMessages = messages.ToList(),
+            };
+            return BadRequest(errors);
+        }
+
+        var artist = ArtistService.GetArtistById(artistId);
+        var album = AlbumService.GetAlbumById(albumId);
         var existingSong = SongService.GetSongById(id);
 
-        if(existingSong is null)
+        if(artist is null || album is null || existingSong is null)
         {
-            return NotFound();
+            return NotFound("Artist or Album or Song is not found");
         }
 
         SongService.UpdateSong(id, song);
@@ -84,11 +152,13 @@ public class SongsController : ControllerBase
     [HttpDelete("{id}")]
     public IActionResult DeleteSong(string artistId, string albumId, string id)
     {
+        var artist = ArtistService.GetArtistById(artistId);
+        var album = AlbumService.GetAlbumById(albumId);
         var song = SongService.GetSongById(id);
 
-        if (song is null)
+        if (artist is null || album is null || song is null)
         {
-            return NotFound();
+            return NotFound("Artist or Album or Song is not found");
         }
 
         SongService.DeleteSong(id);
